CLI: Name ALL the CTAs
I had two options. Either go with 'learn more' for all links, or to come
up with individual texts for every single one of them. I went with the
latter because it's more fun.

I'm not sure about the tone of the buttons. They should probably be less
casual and also in imperative form.