Reword the "no runtime" section
The previous phrasing "or even integrate with other languages" seems to imply (to me) that integrating with other languages is an extraordinary thing. Many languages have FFIs, but where Rust excels is how easy to use its FFI is, not merely that it has one at all. Hence the slight rewording to emphasize this connotation.