Add IRC link back to community page
We have not yet had a community discussion on replacing IRC with Discord. Discord is a proprietary chat platform with a variety of well-documented issues, including accessibility concerns. I believe that it is unacceptable to force this change on the community in order to communicate about and contribute to Rust, as much as both @ashleygwilliams and @steveklabnik have been attempting to sneak this under the table by fending off any criticism with repeated assertions that IRC is not being killed off - even though it evidently is by the removal of the link.

As a result, this change moves Discord into a secondary link, with IRC being the primary link, and this should remain the case until there is a community discussion on this topic.