fix(nav): improve responsive nav behavior.
Define a class to provide flexible widths of the roughly-desirable
size for small screens, and apply that to the nav. Note that it is
not possible to guarantee the list *never* looks odd without CSS Grid,
but here we at least minimize the number of sizes where it looks odd,
and in particular improve its look on all phone views.

Resolves #498.