Cache the current Rust version.
Currently, every time the home page is rendered, the web server
downloads a TOML file from static.rust-lang.org and parses it to find
the current Rust version.  This adds a significant amount of latency to
the home page.

This patch caches the version in memory for up to 10 minutes.  This
greatly speeds up home page rendering, at the cost of a 10-minute delay
before a new version is reflected on the home page.