install-script.hbs: Support case of missing test button
When testing the install page, we can add a button to cycle
the platform divs.  However we don't include that button in
normal use which causes a JS console error.  Since we'd like
the test code to still function in the absence of the button
this tweak allows for that.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>