install-script.hbs: Add support for external override
Sometimes, for example from rustup's documentation, we might
want to link to the install page for a particular platform,
but still have that recommend the usual install methods.

This adds the ability to set `platform_override` via a query
string, also accounting for some people trying to probe options.

For example, we support `?platform_override=generic` to provide
the variant of the instructions which offer both the curl and
the windows exe installation methods.

Fixes: #726

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>