Tweak installation page (#729)
* rustup.hbs: Make generic instructions nicer

The generic installation instructions weren't quite in the same
style as the others, this makes them a little nicer.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>

* install-script.hbs: Support case of missing test button

When testing the install page, we can add a button to cycle
the platform divs.  However we don't include that button in
normal use which causes a JS console error.  Since we'd like
the test code to still function in the absence of the button
this tweak allows for that.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>

* install-script.hbs: Add support for external override

Sometimes, for example from rustup's documentation, we might
want to link to the install page for a particular platform,
but still have that recommend the usual install methods.

This adds the ability to set `platform_override` via a query
string, also accounting for some people trying to probe options.

For example, we support `?platform_override=generic` to provide
the variant of the instructions which offer both the curl and
the windows exe installation methods.

Fixes: #726

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>