Add missing arg for creating new project example
It seems `cargo new` currently defaults to using a library template to create a new project, but the "Generating new project" example assumes it will use the binary template. 

The library template is incompatible with the next steps in the example. When you try to run `cargo run`, you get an error:

```
➜  hello-rust git:(master) ✗ cargo run
error: a bin target must be available for `cargo run`
```

This PR adds the missing `--bin` argument to make the example correct.