Build on Windows with VS2019
This commit contains a temporary fork on `sass-rs` which provides the
necessary chagnes to have the project running on Windows 10.

I expect [the change](https://github.com/compass-rs/sass-rs/pull/49)
to be taken upstream, so we can simply upgrade the dependency instead of
keeping the forked git version, but this is more so to let others
compile the project and document the error, if someone tries to run it
on Windows meanwhile.

The error message in itself, to help with searchs, is:

```
error MSB4019: The imported project "C:\Microsoft.Cpp.Default.props" was
not found. Confirm that the path in the <Import> declaration is correct,
and that the file exists on disk.
```

This also provides a documentation on the README on what to do in case
the error happens.