Allow buttons to grow to accomodate translations
The current buttons make use of a fixed height for the text content,
which make them look similar to each other and provide a pleasant visual
connection.

It happens that some translations require more text space, and the
content of the button started to be left out, specially on mobile
screens as it has less width to use.

This commit changes the buttons to allow them to grow vertically as
necessary, while remaining at a similar height when it is not necessary.

It is easier to see the different using the `Governance` or `Community`
content in Portuguese.

Tested visually both on Desktop and Mobile views, and the content
adjusts to the provided width.

A small extra fix has been included on the CLI page, to make a regular
spacing on the text columns when visiting it on a Mobile view.