Add a privacy notice for all projects managed by Rust teams
This adds a new page containing a privacy notice. This notice includes
sections for rust-lang.org, crates.io, docs.rs, the forums, and other
third party services we use but do not manage (such as Discord). The
intention is that we will link to this page from across the org. The
forums will also link to Discourse's privacy policy.

Since this is a legal document, I'm not certain that we can publish
community translations for this (or at least I'd want to confirm with
lawyers that it's safe for us to do so). As such, in the interest of
time, I've opted to leave the markup inline rather than slicing it up.
If this is a blocker for publication, I can take the time to break it
apart -- but we should get this out sooner rather than later.