Add support for running the page behind a proxy. (#1009)
The rust web page fetch data from the rust-lang.org site
when it loads. If the site is run locally and behind a proxy,
this fetch process will fail with timeout error. `reqwest` needs
to know explicitly the information about the proxy.

This patch add a function to fetch this data by retrieving the
proxy information from the environment variables `https_proxy` and
`HTTPS_PROXY`. Then the `reqwest` client is created accordingly.

A minor refactor was done to call this new function from the
`fetch_rust_*` functions.

Signed-off-by: Erich Cordoba <erich.cm@yandex.com>