Rework issue templates
The current templates currently leave all of the help text in the issue
if the person forgot to manually remove it. This commit puts the text
behind a html comment so that it is absent, as well as reducing the
heading size.

This also reduces the number of templates from seven to three. The
motivation being that a lot of the current templates are unused, and
that the distinction between what it makes it a "content" or "technical"
feature request is not something that the issue writer necessarily knows
about enough to feel confident in saying it's either one.