Re-arrange the community webpage (#1203)
This updates the community webpage to remove the old section on
Increasing Rust Reach and re-arranges the top section of the page to
reduce the prominance of the Community Team's email as currently it
receives a lot of traffic. By moving it below the section on forums this
should hopefully steer people to ask their questions on the forums
instead.