Fix syntax highlighting
Add the `plaintext` language class to some code blocks that were being
highlighted incorrectly. The `cargo run` one was being highlighted as
Ruby and the `ferris_says` one was being highlighted as Bash!

Also added a missing `p` tag around an inline code block that was on its
own line.