policies/privacy: Remove `#privacy-notice` ID from section element
It turns out that certain adblockers (e.g. uBlock Origin) include the "EasyList Cookies" blocklist, which prevents elements with an `#privacy-notice` ID from rendering, which resulted in us showing a blank page. Since this ID does not seem to be used for anything here this commit just removes it to prevent the adblocker issue.