Fix escaping for Zulip streams
This moves the escaping for Zulip streams out of the data pipeline and over to a
Handlebars helper. This in turn allows the group team template to make use of
both the unescaped (for display) and escaped (for the link) variants.

Fixes https://github.com/rust-lang/www.rust-lang.org/issues/1274