Remove link to ping.rust-lang.org from footer
The site has not been up for over a month, and through chatting with
some folks from the infra team on [Zulip], it seems they would prefer to
remove the link rather than trying to restore the site.

I agree that a separate site for status is not awfully useful, as
www.rust-lang.org is mostly if not entirely static content. Simply
visiting the site checks whether it is available.

[Zulip]: https://rust-lang.zulipchat.com/#narrow/stream/242791-t-infra/topic/rust-lang.2Eorg.20.20issues/near/268123956

Closes GH-1619