Fix clippy config
Denying all warnings for test builds results in a poor development
experience, as everying always appears as an error in the editor.
The program might still compile and run fine though.

At the same time, CI only denies rustc's warnings, not those of clippy.
So code that gets marked as an error in an editor can still pass CI.

Both of these issues are fixed by running clippy in CI explicitly.