Fix redirects
This regression was introduced in ca2a1621ece46315bdce0c1f76ed055f3d18a20a.
Redirects like /team.html -> /governance didn't work anymore.

In the update from Rocket v0.5.0-rc.3 to v0.5.0, there was a breaking
change missed in the changelog of v0.5.0-rc.4 that affected us.
https://github.com/rwf2/Rocket/blob/v0.5.0-rc.4/CHANGELOG.md

The relevant section in the changelog starts with the line:
"The status codes used when built-in guards forward were changed."

We used the status code 404 to invoke our custom `not_found` catcher.
This custom catcher is responsible for the redirects.
When the status code changed to 422, this catcher was not invoked and
the redirects were not triggered.

This fix simply invokes the catcher for both status codes 404 and 422.