Remove unrecognized platform report button
There was a steady stream of issues generated by this report button.

None of the people who opened issues would be responsive and eloborate
on the kind of platform they're using. Most of them had "Android" in their
user agent, even though the issue template specifically said not to open
an issue if one is using Android.

At this point, it seems safe to assume the platform detection code
works pretty well and getting rid of the report button is not a big loss.