Remove vestigial check for leadership-council subteam
In https://github.com/rust-lang/team/pull/1403 we ended up doing things
a little differently so that the top-level teams are not subteams of the
council. This removes the check added in
https://github.com/rust-lang/www.rust-lang.org/pull/1937.

This helps fix an issue where we do have some actual subteams of the
council (like social-media), and we don't want those to show up on the
main page.