Add link to Discord server
For Rust users (who are likely the main audience for this page), pointing them to the Zulip server isn't helpful because it's only for development of Rust itself. Therefore I added a link to the Discord "server" (note, Discord uses the term "server" in a stupid way; it's like a Slack workspace).

This is not an official Rust project but it's likely what most users are looking for so adding it here is a pragmatic and useful choice. I described it as "unofficial".

This link is especially helpful because there are two Rust Discover servers, but one of them is defunct (in a way that is annoying to discover). Unfortunately it still tops Google results sometimes depending on what you search for.

Finally, I think the layout could probably be cleaned up by having it split in two:

* For users: [User Forum], [Discord]
* For Rust language developers: [Internals Forum] [Zulip] [Teams]

But that's a bigger change and I haven't done it here.