Invert RBE and Rustlings order
A person reading the Learn Rust page left to right or top to bottom
will read the Rustlings section before the RBE section, but the former
mentions the latter, so the person will encounter RBE before having
read about it.
Inverting the order of the two sections prevents these unnecessary
"forward references".